+++ armada_command/command_restart.py
@@ -1,5 +1,7 @@
 from __future__ import print_function
 import argparse
+import base64
+import json
 import os
 import sys
 import traceback
@@ -58,24 +60,36 @@ def command_restart(args):
             container_id = instance['ServiceID'].split(':')[0]
             is_run_locally = armada_utils.is_local_container(container_id)
 
-            payload = {'container_id': container_id}
-            if args.ship:
-                payload['target_ship'] = args.ship
-                payload['force'] = args.force
-
-            print('Checking if there is new image version. May take few minutes if download is needed...')
-            result = armada_api.post('restart', payload, ship_name=instance['Node'])
-
-            if result['status'] == 'ok':
-                new_container_id = result['container_id']
-                print('Service has been restarted and is running in container {new_container_id} '
-                      'available at addresses:'.format(**locals()))
-                for service_address, docker_port in result['endpoints'].iteritems():
-                    print('  {0} ({1})'.format(service_address, docker_port))
-                if instances_count > 1:
-                    print()
+            if is_run_locally:
+                result = json.loads(armada_api.get('env/{container_id}/ARMADA_RUN_COMMAND'.format(**locals())))
+                if result['status'] == 'ok':
+                    stop_command = 'armada stop {container_id}'.format(**locals())
+                    run_command = base64.b64decode(result['value'])
+                    assert armada_utils.execute_local_command(stop_command, stream_output=True, retries=3)[0] == 0
+                    assert armada_utils.execute_local_command(run_command, stream_output=True, retries=5)[0] == 0
+                    if instances_count > 1:
+                        print()
+                else:
+                    raise armada_utils.ArmadaCommandException(result['error'])
             else:
-                raise armada_utils.ArmadaCommandException(result['error'])
+                payload = {'container_id': container_id}
+                if args.ship:
+                    payload['target_ship'] = args.ship
+                    payload['force'] = args.force
+
+                print('Checking if there is new image version. May take few minutes if download is needed...')
+                result = armada_api.post('restart', payload, ship_name=instance['Node'])
+
+                if result['status'] == 'ok':
+                    new_container_id = result['container_id']
+                    print('Service has been restarted and is running in container {new_container_id} '
+                          'available at addresses:'.format(**locals()))
+                    for service_address, docker_port in result['endpoints'].iteritems():
+                        print('  {0} ({1})'.format(service_address, docker_port))
+                    if instances_count > 1:
+                        print()
+                else:
+                    raise armada_utils.ArmadaCommandException(result['error'])
         except armada_utils.ArmadaCommandException as e:
             print("ArmadaCommandException: {0}".format(str(e)))
             were_errors = True