+++ armada_backend/recover_saved_containers.py
@@ -1,13 +1,12 @@
-from __future__ import print_function
 import argparse
+import json
+import logging
+import os
+import sys
+import traceback
 from collections import Counter
 from time import sleep
-import json
-import os
-import traceback
-import sys
 
-from armada_backend.api_run import print_err
 from armada_backend.api_ship import wait_for_consul_ready
 from armada_backend.utils import get_container_parameters, get_local_containers_ids
 from armada_command import armada_api
@@ -34,9 +33,9 @@ def _get_local_running_containers():
 
 
 def _recover_container(container_parameters):
-    print_err('Recovering: {}...\n'.format(json.dumps(container_parameters)))
+    logging.info('Recovering: {}...\n'.format(json.dumps(container_parameters)))
     recovery_result = armada_api.post('run', container_parameters)
-    print_err('Recovered container: {}'.format(json.dumps(recovery_result)))
+    logging.info('Recovered container: {}'.format(json.dumps(recovery_result)))
 
 
 def _multiset_difference(a, b):
@@ -52,7 +51,7 @@ def recover_saved_containers(saved_containers):
     containers_to_be_recovered = _multiset_difference(saved_containers, running_containers)
     recovery_retry_count = 0
     while containers_to_be_recovered and recovery_retry_count < RECOVERY_RETRY_LIMIT:
-        print_err("Recovering containers: {}".format(json.dumps(containers_to_be_recovered)))
+        logging.info("Recovering containers: {}".format(json.dumps(containers_to_be_recovered)))
         for container_parameters in containers_to_be_recovered:
             _recover_container(container_parameters)
         sleep(DELAY_BETWEEN_RECOVER_RETRY_SECONDS)
@@ -67,29 +66,30 @@ def _recover_saved_containers_from_path(saved_containers_path):
         saved_containers = _load_saved_containers_parameters_list(saved_containers_path)
         not_recovered = recover_saved_containers(saved_containers)
         if not_recovered:
-            print_err('Following containers were not recovered: ', not_recovered)
+            logging.error('Following containers were not recovered: ', not_recovered)
             return False
         else:
             return True
     except:
         traceback.print_exc()
-        print_err('Unable to recover from {saved_containers_path}.'.format(**locals()))
+        logging.error('Unable to recover from {saved_containers_path}.'.format(**locals()))
     return False
 
 
 def _check_if_we_should_recover(saved_containers_path):
     try:
         if int(os.environ.get('DOCKER_START_TIMESTAMP')) > int(os.path.getmtime(saved_containers_path)):
-            print_err('Docker daemon restart detected.')
+            logging.info('Docker daemon restart detected.')
             return True
         else:
-            print_err('No need to recover.')
+            logging.info('No need to recover.')
             return False
     except:
         return False
 
 
 def main():
+    logging.basicConfig(level=logging.INFO)
     try:
         args = _parse_args()
         if args.force or _check_if_we_should_recover(args.saved_containers_path):