+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Hold version of python requests package to 2.9.1 in microservice image.
 - Fixed checking for newest local image of service on restart in vagrant development environment.
 
+
 ## 0.19.0 (2016-05-17)
 
 ### Improvements